crop-disease-backend/requirements.txt
Flask
Flask-Cors
tensorflow
numpy
Pillow
google-generativeai
python-dotenv
